+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/menu/party/GuiMenu_Party_Create.java
@@ -30,7 +30,6 @@ public class GuiMenu_Party_Create extends MenuBackground {
 	MenuButton togglePriv, accept, size;
 	MenuButton back;
 		
-	final PlayerData playerData = PlayerData.get(minecraft.player);
 	WorldData worldData;
 	
 	Party party;
@@ -73,9 +72,7 @@ public class GuiMenu_Party_Create extends MenuBackground {
 	}
 
 	private void updateButtons() {
-		//IPlayerCapabilities playerData = ModCapabilities.getPlayer(minecraft.player);
 		togglePriv.setMessage(priv ? Component.translatable(Utils.translateToLocal(Strings.Gui_Menu_Party_Create_Accessibility_Private)) : Component.translatable(Utils.translateToLocal(Strings.Gui_Menu_Party_Create_Accessibility_Public)));
-
 		
 		//TBName
 		togglePriv.visible = true;