+++ src/main/java/online/kingdomkeys/kingdomkeys/client/gui/menu/party/GuiMenu_Party_Promote.java
@@ -2,7 +2,9 @@ package online.kingdomkeys.kingdomkeys.client.gui.menu.party;
 
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.AbstractWidget;
+import net.minecraft.network.chat.Component;
 import net.minecraft.sounds.SoundSource;
+import online.kingdomkeys.kingdomkeys.client.ClientUtils;
 import online.kingdomkeys.kingdomkeys.client.gui.elements.MenuBackground;
 import online.kingdomkeys.kingdomkeys.client.gui.elements.MenuBox;
 import online.kingdomkeys.kingdomkeys.client.gui.elements.buttons.MenuButton;
@@ -27,7 +29,6 @@ public class GuiMenu_Party_Promote extends MenuBackground {
 	MenuBox box;
 	MenuButton back;
 		
-	final PlayerData playerData = PlayerData.get(minecraft.player);
 	WorldData worldData;
 	Party party;
 	
@@ -145,7 +146,15 @@ public class GuiMenu_Party_Promote extends MenuBackground {
 		super.render(gui, mouseX, mouseY, partialTicks);
 		worldData = WorldData.getClient();
 		party = worldData.getPartyFromMember(minecraft.player.getUUID());
-		refreshMembers(gui);
 
+		gui.pose().pushPose();
+		{
+			float scale = 1.5F;
+			gui.pose().scale(scale, scale, 1);
+			gui.drawString(minecraft.font, Component.literal("["+party.getMembers().size()+"/"+party.getSize()+"] "+party.getName()).withStyle(ClientUtils.KK_Font_EXP), (int) (topLeftBar.getWidth() / scale + topGap) + 5, 10, 0xFF9900);
+		}
+		gui.pose().popPose();
+
+		refreshMembers(gui);
 	}
 }